refactor logo component
