+++ src/components/Logo/Logo.styles.js
@@ -1,0 +1,15 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles(theme => ({
+    iconAvatar: {
+        marginTop: '20px',
+        marginBottom: '10px',
+        marginLeft: 'auto',
+        marginRight: 'auto',
+        width: '100px',
+        height: '100px'
+    },
+    appName: {
+        fontWeight : '600'
+    }
+}));
